packages/pytest-msgspec-schema-sync/pyproject.toml
[build-system]
requires = [
  "pdm-backend"
]
build-backend = "pdm.backend"

[project]
name = "pytest-msgspec-schema-sync"
version = "0.2.1"
description = "Pytest plugin to synchronise msgspec Struct schemas with JSONSchema files"
authors = [
  {name = "Louis Maddox", email = "louismmx@gmail.com"}
]
dependencies = [
  "msgspec-schema-sync>=0.1.5",
  "pytest>=6"
]
requires-python = ">=3.10"
readme = "README.md"

[project.entry-points.pytest11]
msgspec_schema_sync = "pytest_msgspec_schema_sync.plugin"

[project.license]
text = "MIT"

[project.urls]
Homepage = "https://github.com/lmmx/msgspec-schema-sync/packages/pytest-msgspec-schema-sync"
Repository = "https://github.com/lmmx/msgspec-schema-sync.git"

[tool.pdm]
distribution = true

[tool.uv.sources.msgspec-schema-sync]
workspace = true
